ref(metrics): Split MetricSelector into directory with subcomponents and fix side panel styling
Breaks the monolithic MetricSelector into smaller modules:
- types.tsx: MetricSelectOption interface
- metricListBoxOption.tsx: list item component
- metricDetailPanel.tsx: detail panel and attributes section

Also fixes side panel positioning to float alongside the active option
and adapts layout direction when the overlay opens above the trigger.
