+++ static/app/views/explore/metrics/metricToolbar/metricSelector/metricSelector.spec.tsx
@@ -11,7 +11,7 @@ import {
   within,
 } from 'sentry-test/reactTestingLibrary';
 
-import {MetricSelector} from 'sentry/views/explore/metrics/metricToolbar/metricSelector';
+import {MetricSelector} from 'sentry/views/explore/metrics/metricToolbar/metricSelector/metricSelector';
 
 const SORTED_METRIC_NAMES = [
   'bar',
